requirements.txt
flask-cors==3.0.9
Flask
flask-sqlalchemy
flask-mysqldb==0.2.0
mysql==0.0.2 
mysqlclient==2.0.1
sqlalchemy-multidb